global: posts (crud+likes), comment (crud+likes), profil (detail info, comment and post reach, likes reach)

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Post;
+use App\PostLike;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -11,7 +13,12 @@ class PostController extends Controller {
     $this->middleware('auth');
   }
   public function index() {
-    $posts = Post::get();
+    $userId = Auth::user()->id;
+    $posts = Post::with('user', 'likes')
+      ->withCount('likes', 'comments')
+      ->where('user_id', '=', $userId)
+      ->orderBy('id', 'DESC')
+      ->get();
     return view('posts.index', ['posts' => $posts]);
   }
   public function create() {
@@ -38,11 +45,19 @@ class PostController extends Controller {
     }
     Post::create($datas);
     toast('Post Has Been Submitted !', 'success');
-    return redirect('posts');
+    return redirect('/');
   }
 
   public function show(Post $post) {
-
+    $myPost = $post
+      ->with(['comments' => function ($query) {
+        $query->with('likes')->withCount('likes');
+      }])
+      ->with('user', 'likes')
+      ->withCount('likes', 'comments')
+      ->where('id', '=', $post->id)->first();
+    // dd($myPost->toArray());
+    return view('posts.show', ['post' => $myPost]);
   }
 
   public function edit(Post $post) {
@@ -56,6 +71,12 @@ class PostController extends Controller {
       'image' => 'mimes:jpeg,png,jpg,gif,svg',
       'image.*' => 'mimes:jpeg,png,jpg,gif,svg',
     ]);
+    $userId = Auth::user()->id;
+    $userIdPost = $post->user_id;
+    if ($userId != $userIdPost) {
+      toast('This is not your Post', 'error');
+      return redirect('/');
+    }
     $datas = $request->all();
     if ($files = $request->file('image')) {
       // Define upload path
@@ -68,18 +89,37 @@ class PostController extends Controller {
     $post->fill($datas);
     $post->save();
     toast('Post Has Been Updated !', 'success');
-    return redirect('posts');
+    return redirect('/');
   }
 
   public function destroy(Post $post) {
+    $userId = Auth::user()->id;
+    $userIdPost = $post->user_id;
+    if ($userId != $userIdPost) {
+      toast('This is not your Post', 'error');
+      return redirect('/');
+    }
     toast('Post Has Been Deleted !', 'success');
     $post->delete();
-    return redirect('posts');
+    return redirect('/');
   }
 
-  public function like(Post $post) {
+  public function like(Request $request, Post $post) {
     toast('Post Has Been Liked !', 'success');
-    // $post->delete();
-    return redirect('posts');
+    $userId = Auth::user()->id;
+    $datas = $request->all();
+    $datas['user_id'] = $userId;
+    $datas['post_id'] = $post->id;
+    try {
+      PostLike::create($datas);
+    } catch (QueryException $e) {
+      $errorCode = $e->errorInfo[1];
+      if ($errorCode == 1062) {
+        toast('Ups !! Already Liked', 'error');
+        return redirect()->back();
+      }
+    }
+    toast('Comment Has Liked !', 'success');
+    return redirect()->back();
   }
 }
